Optimizations for memory use
Previously, I stored the entire sample of IP addresses in memory until we had
finished sampling from every AS, then we wrote. This works fine for doing more
sophisticated output formats (originally had intended to do kml) or for small
samples (less waiting on disk since everything is in memory) but breaks bad for
samples of meaningful size (4% sample was taking several hours due to swapping).
This commit adds two other mechanisms for large samples:

- Store it all on memory, for small (<1%) sample sizes
- Output as you go and only store a single AS's sample in memory at any given
  time (for most sample sizes 1%-40%)
- Repeatedly perform small samples (0.5%) to reach desired sample size for very
  large samples (i.e., when the sample from a single AS will not fit in memory)

The UI for choosing among these is currently non-existent. The last approach was
quick to write but is probably not optimal for very large samples; fortunately
we don't need anything that large most likely.