Finished collision detection, minor __str__ bug fixes
This collision detection allows for relatively fast generation of the
latlon-to-face lookup table we use to add membership to a particular triangular
grid cell. We model each triangular cell as a sphere; we have a few false
positives but this greatly speeds up collision detection. For each lat/lon
pair, we should only have to check a handful of faces to determine which is the
right one.