UV mapping update, adding face_id for easier parsing/comparisions
The UV mapping code is a little sketchy at this point. UV mapping is essentially
the process of flattening a 3-d object so you can texture map it, which is what
we want to do for drawing bounding boxes. I believe that generating bounding
boxes will be easier done in a flattened image; after we have drawn the bounding
boxes we can re-project the bounding polygons into 3-space and thus map them to
triangles.