Added extent.py, a tool for determining geo extent of an AS from prefix sample
Given a list of <asn> <lat> <lon> pairs sorted by ASN (preferably uniq'd as
well), this performs clustering of points which are less than some threshold
distance away from each other. Given these clusters, we draw a simple bounding
box and call this the AS's extent. The output is a series of latlons that tile
each bounding box (at 0.1 degree intervals) that we can feed to dggrid to get a
list of corresponding cells.