Performance improvements, convex hull cluster growth/shape
All clusters are now considered convex polygons. Rather than growing clusters
by looking for a point that's within the threshold of the new point, we just
check the distance of a new point to the existing bounding hull of the cluster.
If it's inside or within the threshold of the hull's surface, we add it to the
cluster. This makes building clusters 1-2 orders of magnitude faster (run takes
~25min the 5% sample).

Additionally, the "filled" output conforms to the shape of the convex hull,
rather than simply being a square. All points in the cluster are now added to
the filled output.